Author and document datasets

authorData.vicinity_raw() is an authorData method that authors and saves files necessary for creating and documenting datasets. This method is implemented for the ready4 submodule class for tibble object lookup table of metadata about raw (un-processed) spatial data to import. The function returns Files written (a logical vector of length one).

authorData method applied to VicinityLocalRaw

Usage

authorData.vicinity_raw(
  x,
  path_1L_chr,
  data_match_value_xx,
  match_var_nm_1L_chr = "name_chr",
  sub_dirs_chr = NULL,
  overwrite_1L_lgl = F,
  what_1L_chr = "outer"
)

# S4 method for vicinity_raw
authorData(
  x,
  path_1L_chr,
  data_match_value_xx,
  match_var_nm_1L_chr = "name_chr",
  sub_dirs_chr = NULL,
  overwrite_1L_lgl = F,
  what_1L_chr = "outer"
)

# S4 method for VicinityLocalRaw
authorData(x, return_r4_1L_lgl = F)

Arguments

x

An object of class VicinityLocalRaw

path_1L_chr

Path (a character vector of length one)

data_match_value_xx

Data match value (an output object of multiple potential types)

match_var_nm_1L_chr

Match variable name (a character vector of length one), Default: 'name_chr'

sub_dirs_chr

Sub directories (a character vector), Default: NULL

overwrite_1L_lgl

Overwrite (a logical vector of length one), Default: F

what_1L_chr

What (a character vector of length one), Default: 'outer'

return_r4_1L_lgl

Return ready4 module (a logical vector of length one), Default: F

Value

Files written (a logical vector of length one)

Return (an output object of multiple potential types)