Manufacture a new object

manufacture.vicinity_abbreviations() is a manufacture method that manufactures a novel R object using data contained in an instance of a class. This method is implemented for the ready4 submodule class for tibble object lookup table for spatial data abbreviations. The function returns Object (an output object of multiple potential types).

manufacture.vicinity_parameters() is a manufacture method that manufactures a novel R object using data contained in an instance of a class. This method is implemented for the ready4 submodule class for tibble object that stores simulation structural parameters relating to the spatial environment. The function returns Object (an output object of multiple potential types).

manufacture.vicinity_points() is a manufacture method that manufactures a novel R object using data contained in an instance of a class. This method is implemented for the ready4 submodule class for tibble object lookup table of the longitude and latitude cordinates of sites of services / homes. The function returns Object (an output object of multiple potential types).

manufacture.vicinity_processed() is a manufacture method that manufactures a novel R object using data contained in an instance of a class. This method is implemented for the ready4 submodule class for tibble object lookup table of meta-data for spatial data packs (imported and pre-processed data). The function returns Object (an output object of multiple potential types).

manufacture.vicinity_raw() is a manufacture method that manufactures a novel R object using data contained in an instance of a class. This method is implemented for the ready4 submodule class for tibble object lookup table of metadata about raw (un-processed) spatial data to import. The function returns Object (an output object of multiple potential types).

manufacture method applied to VicinityLookup

manufacture method applied to VicinityProfile

Usage

manufacture.vicinity_abbreviations(
  x,
  agent_areas_tb = NULL,
  area_names_chr = character(0),
  area_var_nm_1L_chr = "Suburb",
  areas_sf = NULL,
  correspondences_lup = NULL,
  large_area_var_nm_1L_chr = character(0),
  large_areas_chr = character(0),
  large_area_idx_1L_int = 1L,
  match_value_xx = NULL,
  match_var_nm_1L_chr = character(0),
  outliers_chr = character(0),
  small_area_var_nm_1L_chr = character(0),
  title_case_1L_lgl = T,
  type_1L_chr = "suffix",
  unknown_area_val_1L_chr = "Unk",
  valid_names_chr = character(0),
  what_1L_chr = "areas"
)

# S4 method for vicinity_abbreviations
manufacture(
  x,
  agent_areas_tb = NULL,
  area_names_chr = character(0),
  area_var_nm_1L_chr = "Suburb",
  areas_sf = NULL,
  correspondences_lup = NULL,
  large_area_var_nm_1L_chr = character(0),
  large_areas_chr = character(0),
  large_area_idx_1L_int = 1L,
  match_value_xx = NULL,
  match_var_nm_1L_chr = character(0),
  outliers_chr = character(0),
  small_area_var_nm_1L_chr = character(0),
  title_case_1L_lgl = T,
  type_1L_chr = "suffix",
  unknown_area_val_1L_chr = "Unk",
  valid_names_chr = character(0),
  what_1L_chr = "areas"
)

manufacture.vicinity_parameters(
  x,
  y_vicinity_mapes = NULL,
  n_its_int = integer(0),
  joint_dstr_1L_lgl = T,
  what_1L_chr = "values"
)

# S4 method for vicinity_parameters
manufacture(
  x,
  y_vicinity_mapes = NULL,
  n_its_int = integer(0),
  joint_dstr_1L_lgl = T,
  what_1L_chr = "values"
)

manufacture.vicinity_points(
  x,
  bands_1L_dbl = numeric(0),
  crs_nbr_dbl = numeric(0),
  land_sf = NULL,
  metres_1L_dbl = numeric(0),
  service_1L_chr = character(0),
  time_min_1L_dbl = numeric(0),
  time_max_1L_dbl = numeric(0),
  time_steps_1L_dbl = numeric(0),
  travel_mode_1L_chr = "car",
  type_1L_chr = "single",
  what_1L_chr = "geometric"
)

# S4 method for vicinity_points
manufacture(
  x,
  bands_1L_dbl = numeric(0),
  crs_nbr_dbl = numeric(0),
  land_sf = NULL,
  metres_1L_dbl = numeric(0),
  service_1L_chr = character(0),
  time_min_1L_dbl = numeric(0),
  time_max_1L_dbl = numeric(0),
  time_steps_1L_dbl = numeric(0),
  travel_mode_1L_chr = "car",
  type_1L_chr = "single",
  what_1L_chr = "geometric"
)

manufacture.vicinity_processed(
  x,
  approximation_1L_chr = "abs",
  main_incld_feature_chr = character(0),
  target_area_1L_chr = character(0),
  target_year_1L_chr = character(0),
  what_1L_chr = "closest years"
)

# S4 method for vicinity_processed
manufacture(
  x,
  approximation_1L_chr = "abs",
  main_incld_feature_chr = character(0),
  target_area_1L_chr = character(0),
  target_year_1L_chr = character(0),
  what_1L_chr = "closest years"
)

manufacture.vicinity_raw(
  x,
  args_ls = NULL,
  crs_nbr_dbl = NA_real_,
  fn = function(x, ...) {
     NULL
 },
  match_value_xx = NULL,
  match_var_nm_1L_chr = character(0),
  merge_itms_chr = character(0),
  overwrite_1L_lgl = F,
  package_1L_chr = character(0),
  processed_fls_dir_1L_chr = character(0),
  raw_fls_dir_1L_chr = character(0),
  sub_dirs_chr = character(0),
  what_1L_chr = "output"
)

# S4 method for vicinity_raw
manufacture(
  x,
  args_ls = NULL,
  crs_nbr_dbl = NA_real_,
  fn = function(x, ...) {
     NULL
 },
  match_value_xx = NULL,
  match_var_nm_1L_chr = character(0),
  merge_itms_chr = character(0),
  overwrite_1L_lgl = F,
  package_1L_chr = character(0),
  processed_fls_dir_1L_chr = character(0),
  raw_fls_dir_1L_chr = character(0),
  sub_dirs_chr = character(0),
  what_1L_chr = "output"
)

# S4 method for VicinityLookup
manufacture(
  x,
  area_sf = NULL,
  area_unit_1L_chr = character(0),
  att_data_xx = NULL,
  boundary_year_1L_chr = character(0),
  match_value_xx = character(0),
  path_1L_chr = character(0),
  type_1L_chr = "Geometry",
  what_1L_chr = "attribute",
  y_vicinity_raw = NULL,
  ...
)

# S4 method for VicinityProfile
manufacture(
  x,
  attributes_ls = NULL,
  attributes_to_import_chr = character(0),
  exclude_dif_bndy_yr_1L_lgl = TRUE,
  highest_rsl_chr = character(0),
  key_var_1L_chr = character(0),
  match_year_1L_lgl = FALSE,
  model_start_ymdhms_dtm = NULL,
  nbr_steps_start_to_end_1L_int = integer(0),
  reference_var_rsl_1L_chr = NULL,
  reference_vals_chr = c("tot_pop", "age_sex"),
  simulation_steps_ymwd_dtm = NULL,
  specified_rsl_chr = character(0),
  subdivision_1L_chr = NULL,
  type_1L_chr = "inner",
  what_1L_chr = "attributes",
  years_chr = character(0)
)

Arguments

x

An object of class VicinityProfile

agent_areas_tb

Agent areas (a tibble), Default: NULL

area_names_chr

Area names (a character vector), Default: character(0)

area_var_nm_1L_chr

Area variable name (a character vector of length one), Default: 'Suburb'

areas_sf

Areas (a simple features object), Default: NULL

correspondences_lup

Correspondences (a lookup table), Default: NULL

large_area_var_nm_1L_chr

Large area variable name (a character vector of length one), Default: character(0)

large_areas_chr

Large areas (a character vector), Default: character(0)

large_area_idx_1L_int

Large area index (an integer vector of length one), Default: 1

match_value_xx

Match value (an output object of multiple potential types), Default: character(0)

match_var_nm_1L_chr

Match variable name (a character vector of length one), Default: character(0)

outliers_chr

Outliers (a character vector), Default: character(0)

small_area_var_nm_1L_chr

Small area variable name (a character vector of length one), Default: character(0)

title_case_1L_lgl

Title case (a logical vector of length one), Default: T

type_1L_chr

Type (a character vector of length one), Default: 'inner'

unknown_area_val_1L_chr

Unknown area value (a character vector of length one), Default: 'Unk'

valid_names_chr

Valid names (a character vector), Default: character(0)

what_1L_chr

What (a character vector of length one), Default: 'attributes'

y_vicinity_mapes

PARAM_DESCRIPTION, Default: NULL

n_its_int

N its (an integer vector), Default: integer(0)

joint_dstr_1L_lgl

Joint distribution (a logical vector of length one), Default: T

bands_1L_dbl

Bands (a double vector of length one), Default: numeric(0)

crs_nbr_dbl

Coordinates reference system number (a double vector), Default: NA

land_sf

Land (a simple features object), Default: NULL

metres_1L_dbl

Metres (a double vector of length one), Default: numeric(0)

service_1L_chr

Service (a character vector of length one), Default: character(0)

time_min_1L_dbl

Time minimum (a double vector of length one), Default: numeric(0)

time_max_1L_dbl

Time maximum (a double vector of length one), Default: numeric(0)

time_steps_1L_dbl

Time steps (a double vector of length one), Default: numeric(0)

travel_mode_1L_chr

Travel mode (a character vector of length one), Default: 'car'

approximation_1L_chr

Approximation (a character vector of length one), Default: 'abs'

main_incld_feature_chr

Main included feature (a character vector), Default: character(0)

target_area_1L_chr

Target area (a character vector of length one), Default: character(0)

target_year_1L_chr

Target year (a character vector of length one), Default: character(0)

args_ls

Arguments (a list), Default: NULL

fn

Function (a function), Default: function(x, ...) NULL

merge_itms_chr

Merge items (a character vector), Default: character(0)

overwrite_1L_lgl

Overwrite (a logical vector of length one), Default: F

package_1L_chr

Package (a character vector of length one), Default: character(0)

processed_fls_dir_1L_chr

Processed files directory (a character vector of length one), Default: character(0)

raw_fls_dir_1L_chr

Raw files directory (a character vector of length one), Default: character(0)

sub_dirs_chr

Sub directories (a character vector), Default: character(0)

area_sf

Area (a simple features object), Default: NULL

area_unit_1L_chr

Area unit (a character vector of length one), Default: character(0)

att_data_xx

Attribute data (an output object of multiple potential types), Default: NULL

boundary_year_1L_chr

Boundary year (a character vector of length one), Default: character(0)

path_1L_chr

Path (a character vector of length one), Default: character(0)

y_vicinity_raw

PARAM_DESCRIPTION, Default: NULL

...

Additional arguments

attributes_ls

Attributes (a list), Default: NULL

attributes_to_import_chr

Attributes to import (a character vector), Default: character(0)

exclude_dif_bndy_yr_1L_lgl

Exclude different boundary year (a logical vector of length one), Default: TRUE

highest_rsl_chr

Highest resolution (a character vector), Default: character(0)

key_var_1L_chr

Key variable (a character vector of length one), Default: character(0)

match_year_1L_lgl

Match year (a logical vector of length one), Default: FALSE

model_start_ymdhms_dtm

Model start years months days hours minutes seconds (a date vector), Default: NULL

nbr_steps_start_to_end_1L_int

Number steps start to end (an integer vector of length one), Default: integer(0)

reference_var_rsl_1L_chr

Reference variable resolution (a character vector of length one), Default: NULL

reference_vals_chr

Reference values (a character vector), Default: c("tot_pop", "age_sex")

simulation_steps_ymwd_dtm

Simulation steps years month weeks days (a date vector), Default: NULL

specified_rsl_chr

Specified resolution (a character vector), Default: character(0)

subdivision_1L_chr

Subdivision (a character vector of length one), Default: NULL

years_chr

Years (a character vector), Default: character(0)

Value

Object (an output object of multiple potential types)

Object (an output object of multiple potential types)

Object (an output object of multiple potential types)

Object (an output object of multiple potential types)

Object (an output object of multiple potential types)

Object (an output object of multiple potential types)

Object (an output object of multiple potential types)