Procure items from a dataset

procure.vicinity_abbreviations() is a procure method that procures data by executing a search and retrieval algorithm using data contained in an instance of a class. This method is implemented for the ready4 S3 class for tibble object lookup table for spatial data abbreviations. The function returns Object (an output object of multiple potential types).

procure.vicinity_identifiers() is a procure method that procures data by executing a search and retrieval algorithm using data contained in an instance of a class. This method is implemented for the ready4 S3 class for tibble object lookup table of unique feature identifiers used for different spatial objects. The function returns Object (an output object of multiple potential types).

procure.vicinity_raw() is a procure method that procures data by executing a search and retrieval algorithm using data contained in an instance of a class. This method is implemented for the ready4 S3 class for tibble object lookup table of metadata about raw (un-processed) spatial data to import. The function returns Object (an output object of multiple potential types).

procure.vicinity_resolutions() is a procure method that procures data by executing a search and retrieval algorithm using data contained in an instance of a class. This method is implemented for the ready4 S3 class for tibble object lookup table of the relative resolutions of different spatial objects. The function returns Resolution (an output object of multiple potential types).

procure method applied to VicinityProfile

Usage

procure.vicinity_abbreviations(
  x,
  col_nm_1L_chr = "short_name_chr",
  match_value_xx
)

# S4 method for vicinity_abbreviations
procure(x, col_nm_1L_chr = "short_name_chr", match_value_xx)

procure.vicinity_identifiers(
  x,
  col_nm_1L_chr = "spatial_unit_chr",
  geometry_rsl_1L_chr = character(0),
  group_at_geom_unit_1L_lgl = TRUE,
  data_rsl_1L_chr = character(0),
  match_value_xx = NULL,
  area_bndy_yr_chr = character(0),
  what_1L_chr = "match"
)

# S4 method for vicinity_identifiers
procure(
  x,
  col_nm_1L_chr = "spatial_unit_chr",
  geometry_rsl_1L_chr = character(0),
  group_at_geom_unit_1L_lgl = TRUE,
  data_rsl_1L_chr = character(0),
  match_value_xx = NULL,
  area_bndy_yr_chr = character(0),
  what_1L_chr = "match"
)

procure.vicinity_raw(
  x,
  inc_script_1L_lgl = T,
  forced_choice_chr = NA_character_,
  match_value_xx = NULL,
  what_1L_chr = "source"
)

# S4 method for vicinity_raw
procure(
  x,
  inc_script_1L_lgl = T,
  forced_choice_chr = NA_character_,
  match_value_xx = NULL,
  what_1L_chr = "source"
)

procure.vicinity_resolutions(
  x,
  options_chr = character(0),
  what_1L_chr = "resolution",
  whole_area_1L_lgl = TRUE,
  year_1L_dbl = numeric(0)
)

# S4 method for vicinity_resolutions
procure(
  x,
  options_chr = character(0),
  what_1L_chr = "resolution",
  whole_area_1L_lgl = TRUE,
  year_1L_dbl = numeric(0)
)

# S4 method for VicinityProfile
procure(
  x,
  exclude_dif_bndy_yr_1L_lgl = TRUE,
  highest_rsl_chr = character(0),
  key_var_1L_chr = character(0),
  match_year_1L_lgl = TRUE,
  travel_mode_1L_chr = character(0),
  years_chr = character(0),
  what_1L_chr = "grouping"
)

Arguments

x

An object of class VicinityProfile

col_nm_1L_chr

Column name (a character vector of length one), Default: 'spatial_unit_chr'

match_value_xx

Match value (an output object of multiple potential types), Default: NULL

geometry_rsl_1L_chr

Geometry resolution (a character vector of length one), Default: character(0)

group_at_geom_unit_1L_lgl

Group at geometry unit (a logical vector of length one), Default: TRUE

data_rsl_1L_chr

Data resolution (a character vector of length one), Default: character(0)

area_bndy_yr_chr

Area boundary year (a character vector), Default: character(0)

what_1L_chr

What (a character vector of length one), Default: 'grouping'

inc_script_1L_lgl

Include script (a logical vector of length one), Default: T

forced_choice_chr

Forced choice (a character vector), Default: 'NA'

options_chr

Options (a character vector), Default: character(0)

whole_area_1L_lgl

Whole area (a logical vector of length one), Default: TRUE

year_1L_dbl

Year (a double vector of length one), Default: numeric(0)

exclude_dif_bndy_yr_1L_lgl

Exclude different boundary year (a logical vector of length one), Default: TRUE

highest_rsl_chr

Highest resolution (a character vector), Default: character(0)

key_var_1L_chr

Key variable (a character vector of length one), Default: character(0)

match_year_1L_lgl

Match year (a logical vector of length one), Default: TRUE

travel_mode_1L_chr

Travel mode (a character vector of length one), Default: character(0)

years_chr

Years (a character vector), Default: character(0)

Value

Object (an output object of multiple potential types)

Object (an output object of multiple potential types)

Object (an output object of multiple potential types)

Resolution (an output object of multiple potential types)

Object (an output object of multiple potential types)